+++ services/src/test/java/io/grpc/services/internal/CensusBinaryLogProviderTest.java
@@ -19,10 +19,10 @@ package io.grpc.services.internal;
 import static com.google.common.truth.Truth.assertThat;
 import static io.opencensus.trace.unsafe.ContextUtils.CONTEXT_SPAN_KEY;
 
+import io.grpc.BinaryLogProvider;
+import io.grpc.BinaryLogProvider.CallId;
 import io.grpc.CallOptions;
 import io.grpc.Context;
-import io.grpc.internal.BinaryLogProvider;
-import io.grpc.internal.BinaryLogProvider.CallId;
 import io.grpc.internal.testing.StatsTestUtils.MockableSpan;
 import java.nio.ByteBuffer;
 import java.util.Random;
